[#2149] - Separa la cara de LiteraryWork de los mocks de dominio
Cada archivo del corpus exportaba dos entidades a la vez: la story y la
obra literaria del mismo slug, más sus epígrafes donde los hay. Ninguna
carpeta podía alojarlo entero, así que agrupar por entidad exigía partirlo
primero.

Al separarlas queda a la vista una dependencia que hasta ahora era
invisible por vivir en el mismo archivo: la obra deriva de la story siete
campos —slug, título, autor, portada, recursos, publicación original y
fecha—, y ahora eso es un import explícito. Que se vea es el punto: era
una relación real que el archivo único escondía.
