+++ src/mocks/onoff-literary-works.mock.ts
@@ -1,13 +1,13 @@
 import type { LiteraryWork } from '@models/literary-work.model';
 import type { AttributedText } from '@models/attributed-text.model';
-import { elOdioLiteraryWorkMock } from './onoff/el-odio.mock';
-import { elTratadoDeLosPlaceresLiteraryWorkMock } from './onoff/el-tratado-de-los-placeres.mock';
-import { geometriaLiteraryWorkMock } from './onoff/geometria.mock';
-import { lasDosAntorchasLiteraryWorkMock } from './onoff/las-dos-antorchas.mock';
-import { lasEscalerasLiteraryWorkMock } from './onoff/las-escaleras.mock';
-import { losPeldanosLiteraryWorkMock } from './onoff/los-peldanos.mock';
-import { neronLiteraryWorkMock } from './onoff/neron.mock';
-import { palacioNueveFronterasLiteraryWorkMock } from './onoff/el-palacio-de-las-nueve-fronteras.mock';
+import { elOdioLiteraryWorkMock } from './onoff/literary-work/el-odio.literary-work.mock';
+import { elTratadoDeLosPlaceresLiteraryWorkMock } from './onoff/literary-work/el-tratado-de-los-placeres.literary-work.mock';
+import { geometriaLiteraryWorkMock } from './onoff/literary-work/geometria.literary-work.mock';
+import { lasDosAntorchasLiteraryWorkMock } from './onoff/literary-work/las-dos-antorchas.literary-work.mock';
+import { lasEscalerasLiteraryWorkMock } from './onoff/literary-work/las-escaleras.literary-work.mock';
+import { losPeldanosLiteraryWorkMock } from './onoff/literary-work/los-peldanos.literary-work.mock';
+import { neronLiteraryWorkMock } from './onoff/literary-work/neron.literary-work.mock';
+import { palacioNueveFronterasLiteraryWorkMock } from './onoff/literary-work/el-palacio-de-las-nueve-fronteras.literary-work.mock';
 
 // Corpus LiteraryWork de las obras (ficticias) de François Onoff — contraparte del corpus Story, con
 // el cuerpo en Markdown (./onoff/<slug>.md) materializado a bodyHtml por el pipeline del dominio.
