+++ src/mocks/onoff/el-palacio-de-las-nueve-fronteras.mock.ts
@@ -1,21 +1,5 @@
 import type { Story } from '@models/story.model';
 import { elPalacioRawStory } from './story/el-palacio-de-las-nueve-fronteras.story.raw.mock';
-import { palacioNueveFronterasRawLiteraryWork } from './literary-work/el-palacio-de-las-nueve-fronteras.literary-work.raw.mock';
-import { createLiteraryWork, type LiteraryWork } from '@models/literary-work.model';
-import { createAttributedText } from '@models/attributed-text.model';
-import { createLiteraryWorkSection } from '@models/literary-work-section.model';
-import { createSectionTitle } from '@models/section-title.model';
-import { createMarkdown } from '@models/markdown.model';
-import { deriveSectionReadingTime } from '@models/reading-time.model';
-import { createIsoDateTime } from '@utils/date.utils';
-import { markdownToSanitizedHtml } from '@utils/markdown-pipeline.utils';
-import palacioNueveFronterasMdBody from './literary-work/el-palacio-de-las-nueve-fronteras.md?raw';
-import elPalacioDeLasNueveFronterasEditorialNoteMd from './literary-work/el-palacio-de-las-nueve-fronteras.editorial-note.md?raw';
-import {
-	palacioNueveFronterasEpigraphReference,
-	palacioNueveFronterasEpigraphText,
-	palacioNueveFronterasSectionTitle,
-} from './literary-work/el-palacio-de-las-nueve-fronteras.epigraph';
 import { authorMock } from '../author.mock';
 import { toDomainTags } from '../onoff-tags.mock';
 
@@ -267,34 +251,3 @@ export const palacioNueveFronterasStoryMock: Story = {
 		},
 	],
 };
-
-const palacioNueveFronterasBody = createMarkdown(palacioNueveFronterasMdBody);
-
-export const palacioNueveFronterasEpigraphMock = createAttributedText({
-	text: markdownToSanitizedHtml(createMarkdown(palacioNueveFronterasEpigraphText)),
-	reference: markdownToSanitizedHtml(createMarkdown(palacioNueveFronterasEpigraphReference)),
-});
-
-export const palacioNueveFronterasLiteraryWorkMock: LiteraryWork = createLiteraryWork({
-	_id: 'onoff-literary-work-el-palacio-de-las-nueve-fronteras',
-	slug: palacioNueveFronterasStoryMock.slug,
-	title: palacioNueveFronterasStoryMock.title,
-	authors: [palacioNueveFronterasStoryMock.author],
-	coverImage: palacioNueveFronterasStoryMock.coverImage,
-	content: [
-		createLiteraryWorkSection({
-			position: 0,
-			title: createSectionTitle(palacioNueveFronterasSectionTitle),
-			epigraphs: [palacioNueveFronterasEpigraphMock],
-			bodyHtml: markdownToSanitizedHtml(palacioNueveFronterasBody),
-			readingTime: deriveSectionReadingTime(palacioNueveFronterasBody),
-		}),
-	],
-	mediaSources: [],
-	resources: palacioNueveFronterasStoryMock.resources,
-	badLanguage: palacioNueveFronterasStoryMock.badLanguage,
-	tags: toDomainTags(palacioNueveFronterasRawLiteraryWork.tags),
-	originalPublication: palacioNueveFronterasStoryMock.originalPublication,
-	editorialNote: markdownToSanitizedHtml(createMarkdown(elPalacioDeLasNueveFronterasEditorialNoteMd)),
-	publishedAt: createIsoDateTime(palacioNueveFronterasStoryMock.publishedAt),
-});
