+++ src/mocks/onoff/el-tratado-de-los-placeres.mock.ts
@@ -1,14 +1,5 @@
 import type { Story } from '@models/story.model';
 import { elTratadoRawStory } from './story/el-tratado-de-los-placeres.story.raw.mock';
-import { elTratadoDeLosPlaceresRawLiteraryWork } from './literary-work/el-tratado-de-los-placeres.literary-work.raw.mock';
-import { createLiteraryWork, type LiteraryWork } from '@models/literary-work.model';
-import { createLiteraryWorkSection } from '@models/literary-work-section.model';
-import { createMarkdown } from '@models/markdown.model';
-import { deriveSectionReadingTime } from '@models/reading-time.model';
-import { createIsoDateTime } from '@utils/date.utils';
-import { markdownToSanitizedHtml } from '@utils/markdown-pipeline.utils';
-import elTratadoDeLosPlaceresMdBody from './literary-work/el-tratado-de-los-placeres.md?raw';
-import elTratadoDeLosPlaceresEditorialNoteMd from './literary-work/el-tratado-de-los-placeres.editorial-note.md?raw';
 import { authorMock } from '../author.mock';
 import { toDomainTags } from '../onoff-tags.mock';
 
@@ -282,27 +273,3 @@ export const elTratadoDeLosPlaceresStoryMock: Story = {
 		},
 	],
 };
-
-const elTratadoDeLosPlaceresBody = createMarkdown(elTratadoDeLosPlaceresMdBody);
-
-export const elTratadoDeLosPlaceresLiteraryWorkMock: LiteraryWork = createLiteraryWork({
-	_id: 'onoff-literary-work-el-tratado-de-los-placeres',
-	slug: elTratadoDeLosPlaceresStoryMock.slug,
-	title: elTratadoDeLosPlaceresStoryMock.title,
-	authors: [elTratadoDeLosPlaceresStoryMock.author],
-	coverImage: elTratadoDeLosPlaceresStoryMock.coverImage,
-	content: [
-		createLiteraryWorkSection({
-			position: 0,
-			bodyHtml: markdownToSanitizedHtml(elTratadoDeLosPlaceresBody),
-			readingTime: deriveSectionReadingTime(elTratadoDeLosPlaceresBody),
-		}),
-	],
-	mediaSources: [],
-	resources: elTratadoDeLosPlaceresStoryMock.resources,
-	badLanguage: elTratadoDeLosPlaceresStoryMock.badLanguage,
-	tags: toDomainTags(elTratadoDeLosPlaceresRawLiteraryWork.tags),
-	originalPublication: elTratadoDeLosPlaceresStoryMock.originalPublication,
-	editorialNote: markdownToSanitizedHtml(createMarkdown(elTratadoDeLosPlaceresEditorialNoteMd)),
-	publishedAt: createIsoDateTime(elTratadoDeLosPlaceresStoryMock.publishedAt),
-});
