+++ src/mocks/onoff/los-peldanos.mock.ts
@@ -1,14 +1,5 @@
 import type { Story } from '@models/story.model';
 import { losPeldanosRawStory } from './story/los-peldanos.story.raw.mock';
-import { losPeldanosRawLiteraryWork } from './literary-work/los-peldanos.literary-work.raw.mock';
-import { createLiteraryWork, type LiteraryWork } from '@models/literary-work.model';
-import { createLiteraryWorkSection } from '@models/literary-work-section.model';
-import { createMarkdown } from '@models/markdown.model';
-import { deriveSectionReadingTime } from '@models/reading-time.model';
-import { createIsoDateTime } from '@utils/date.utils';
-import { markdownToSanitizedHtml } from '@utils/markdown-pipeline.utils';
-import losPeldanosMdBody from './literary-work/los-peldanos.md?raw';
-import losPeldanosEditorialNoteMd from './literary-work/los-peldanos.editorial-note.md?raw';
 import { authorMock } from '../author.mock';
 import { toDomainTags } from '../onoff-tags.mock';
 
@@ -270,27 +261,3 @@ export const losPeldanosStoryMock: Story = {
 		},
 	],
 };
-
-const losPeldanosBody = createMarkdown(losPeldanosMdBody);
-
-export const losPeldanosLiteraryWorkMock: LiteraryWork = createLiteraryWork({
-	_id: 'onoff-literary-work-los-peldanos',
-	slug: losPeldanosStoryMock.slug,
-	title: losPeldanosStoryMock.title,
-	authors: [losPeldanosStoryMock.author],
-	coverImage: losPeldanosStoryMock.coverImage,
-	content: [
-		createLiteraryWorkSection({
-			position: 0,
-			bodyHtml: markdownToSanitizedHtml(losPeldanosBody),
-			readingTime: deriveSectionReadingTime(losPeldanosBody),
-		}),
-	],
-	mediaSources: [],
-	resources: losPeldanosStoryMock.resources,
-	badLanguage: losPeldanosStoryMock.badLanguage,
-	tags: toDomainTags(losPeldanosRawLiteraryWork.tags),
-	originalPublication: losPeldanosStoryMock.originalPublication,
-	editorialNote: markdownToSanitizedHtml(createMarkdown(losPeldanosEditorialNoteMd)),
-	publishedAt: createIsoDateTime(losPeldanosStoryMock.publishedAt),
-});
