+++ src/mocks/onoff/neron.mock.ts
@@ -1,13 +1,5 @@
 import type { Story } from '@models/story.model';
 import { neronRawStory } from './story/neron.story.raw.mock';
-import { neronRawLiteraryWork } from './literary-work/neron.literary-work.raw.mock';
-import { createLiteraryWork, type LiteraryWork } from '@models/literary-work.model';
-import { createLiteraryWorkSection } from '@models/literary-work-section.model';
-import { createMarkdown } from '@models/markdown.model';
-import { deriveSectionReadingTime } from '@models/reading-time.model';
-import { createIsoDateTime } from '@utils/date.utils';
-import { markdownToSanitizedHtml } from '@utils/markdown-pipeline.utils';
-import neronMdBody from './literary-work/neron.md?raw';
 import { authorMock } from '../author.mock';
 import { toDomainTags } from '../onoff-tags.mock';
 
@@ -292,26 +284,3 @@ export const neronStoryMock: Story = {
 		},
 	],
 };
-
-const neronBody = createMarkdown(neronMdBody);
-
-export const neronLiteraryWorkMock: LiteraryWork = createLiteraryWork({
-	_id: 'onoff-literary-work-neron',
-	slug: neronStoryMock.slug,
-	title: neronStoryMock.title,
-	authors: [neronStoryMock.author],
-	coverImage: neronStoryMock.coverImage,
-	content: [
-		createLiteraryWorkSection({
-			position: 0,
-			bodyHtml: markdownToSanitizedHtml(neronBody),
-			readingTime: deriveSectionReadingTime(neronBody),
-		}),
-	],
-	mediaSources: [],
-	resources: neronStoryMock.resources,
-	badLanguage: neronStoryMock.badLanguage,
-	tags: toDomainTags(neronRawLiteraryWork.tags),
-	originalPublication: neronStoryMock.originalPublication,
-	publishedAt: createIsoDateTime(neronStoryMock.publishedAt),
-});
